requirements.txt
cloudpickle
hydra-core
omegaconf==2.1.0
pybind11
flake8==3.8.1 
isort==4.3.21
black==22.3.0
autoflake
timm
pytest
scipy==1.7.3
psutil
opencv-python
wandb
submitit
shapely
pycocoevalcap
einops
torchtext
